added native look editor and the set info panel
git-svn-id: svn://svn.code.sf.net/p/magicseteditor/code/trunk@82 0fc631ac-6414-0410-93d0-97cfa31319b6

+++ src/gui/control/card_list.cpp
@@ -29,7 +29,7 @@ DEFINE_EVENT_TYPE(EVENT_CARD_SELECT);
 
 // ----------------------------------------------------------------------------- : CardListBase
 
-CardListBase::CardListBase(Window* parent, int id, int additional_style)
+CardListBase::CardListBase(Window* parent, int id, long additional_style)
 	: wxListView(parent, id, wxDefaultPosition, wxDefaultSize, additional_style | wxLC_REPORT | wxLC_VIRTUAL | wxLC_SINGLE_SEL)
 {
 	// create image list
